special case fw.to
fw.to doesn't send a 3XX redirect with a Location header to when the client
sends a User-Agent indicating that it is a browser. Instead they send a
META refresh and a JavaScript redirect. This causes a problem for
unshrtn since unshrtn doesn't actually render HTML and execute
JavaScript when the response is 200 OK.

Unfortunately unshrtn added a Chrome's User-Agent because some shortening
services don't like redirecting what don't appear to be browsers.

So this commit special cases fw.to (and potentially others as they are
discovered) which don't redirect browsers with a HTTP redirect.

Fixes #8