metadata
With this commit unshrtn returns canonical links separately from the long url
in case a client wants to distinguish between the two. In addition the title
from the HTML page is also returned, along with the HTTP status andj
content type headers from the final HTTP response.

Internally the leveldb datbase now stores a serialized JSON record as the value,
whereas previously the value was just a URL.

This should set the stage for storing other things such as Twitter/Facebook
metadata. But perhaps the project should be renamed to something new then since
it will be doing a lot more than unshortening URLs. But I still kinda like the
name even if it doens't make any sense.